+++ libpdp/src/sepdp/sepdp_file.c
@@ -38,7 +38,7 @@ int sepdp_write_tags_to_file(const pdp_ctx_t *ctx, const pdp_sepdp_tagdata_t* t)
 {
     unsigned char *buf = NULL;
     FILE *tagfile = NULL;
-    size_t len = 0;
+    unsigned int len = 0;
     size_t off = 0;
     int status = -1;
 
@@ -89,7 +89,7 @@ int sepdp_get_tag_file(const pdp_ctx_t* ctx, unsigned int index,
     pdp_sepdp_tag_t *tag = NULL;
     static FILE* tagfile = NULL;
     unsigned char *data = NULL;
-    size_t data_len = 0;
+    unsigned int data_len = 0;
     int status = -1;
     
     // If buf is NULL, it is a signal to re-set the stateful variables
