+++ libpdp/src/sepdp/sepdp_s3.c
@@ -39,7 +39,7 @@ int sepdp_write_data_to_s3(const pdp_ctx_t *ctx,
     return -1;
 #else
     unsigned char *buf = NULL;
-    size_t len = 0;
+    unsigned int len = 0;
     int status = -1;
 
     if (!t || !t->tokens) return -1;
@@ -86,8 +86,8 @@ int sepdp_get_tag_s3(const pdp_ctx_t* ctx, unsigned int index,
     pdp_sepdp_tag_t *tag = NULL;
     unsigned char *buf = NULL;
     unsigned int i = 0;
-    size_t buf_len = 0;
-    size_t buf_size = 0;
+    unsigned int buf_len = 0;
+    unsigned int buf_size = 0;
     int status = -1;
 
     if (!ctx || !tag_ptr)
